Fix guest Prot::READ permission comparison width
unicorn-engine 2.1.5 types MemRegion::perms as u32 while Prot::READ.0 is
i32, so the two readability checks added by #1407 and #1413 could not
compile. Every other permission test in this file, and the
guest_range_has_permission helper, already casts through `as u32`; these
two sites were the outliers.

Closes #1456

+++ guest/crates/aex-guest-worker/src/x64/imports.rs
@@ -4987,7 +4987,7 @@ fn emulate_output_debug_string_a(unicorn: &mut Unicorn<'_, GuestState>) {
                 .find(|region| {
                     region.begin <= cursor
                         && cursor <= region.end
-                        && region.perms & Prot::READ.0 == Prot::READ.0
+                        && region.perms & Prot::READ.0 as u32 == Prot::READ.0 as u32
                 })
                 .ok_or_else(|| {
                     format!("OutputDebugStringA string at {cursor:#x} is unreadable")
@@ -5210,7 +5210,7 @@ fn read_strncpy_s_source(
             .find(|region| {
                 region.begin <= cursor
                     && cursor <= region.end
-                    && region.perms & Prot::READ.0 == Prot::READ.0
+                    && region.perms & Prot::READ.0 as u32 == Prot::READ.0 as u32
             })
             .ok_or_else(|| format!("strncpy_s source at {cursor:#x} is not readable"))?;
         let available = region
